Add script profile <picture> (<source>) optimisation
