+++ src/styles/layout/sections/_about.scss
@@ -1,3 +1,8 @@
 .about {
     max-width: 900px;
+    &__container{
+        display: grid;
+        grid-template-columns: 3fr 2fr;
+        gap: 50px;
+    }
 }
